Clean the solution, update packages, remove NuGet restore and add Owin.StatelessAuth.Jwt
