+++ samples/Owin.StatelessAuth.Sample/Content/angular-localForage.js
