+++ samples/Owin.StatelessAuth.Sample/Content/app.js
