+++ samples/Owin.StatelessAuth.Sample/Content/localforage.js
