+++ samples/Owin.StatelessAuth.Sample/Views/login.html
