--- src/Owin.StatelessAuth.sln
@@ -1,52 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-VisualStudioVersion = 12.0.30501.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Owin.StatelessAuth", "Owin.StatelessAuth\Owin.StatelessAuth.csproj", "{F8CF626C-D9B6-43A1-B09E-3C44BD748D6C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Owin.StatelessAuth.Tests", "Owin.StatelessAuth.Tests\Owin.StatelessAuth.Tests.csproj", "{FF5C2AA7-0727-4AA9-98E7-23B707533026}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{E5540E85-50F8-441C-96A8-D2656FDF5AF6}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Owin.StatelessAuthExample", "Owin.StatelessAuthExample\Owin.StatelessAuthExample.csproj", "{6A898BCF-E7A8-43CB-B80B-9E7CB2BBC572}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Owin.StatelessAuthExample.Tests", "Owin.StatelessAuthExample.Tests\Owin.StatelessAuthExample.Tests.csproj", "{F30F3427-9A3A-4F1C-88BB-BBD34638AF40}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6A898BCF-E7A8-43CB-B80B-9E7CB2BBC572}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6A898BCF-E7A8-43CB-B80B-9E7CB2BBC572}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6A898BCF-E7A8-43CB-B80B-9E7CB2BBC572}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6A898BCF-E7A8-43CB-B80B-9E7CB2BBC572}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F30F3427-9A3A-4F1C-88BB-BBD34638AF40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F30F3427-9A3A-4F1C-88BB-BBD34638AF40}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F30F3427-9A3A-4F1C-88BB-BBD34638AF40}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F30F3427-9A3A-4F1C-88BB-BBD34638AF40}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F8CF626C-D9B6-43A1-B09E-3C44BD748D6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F8CF626C-D9B6-43A1-B09E-3C44BD748D6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8CF626C-D9B6-43A1-B09E-3C44BD748D6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F8CF626C-D9B6-43A1-B09E-3C44BD748D6C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FF5C2AA7-0727-4AA9-98E7-23B707533026}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FF5C2AA7-0727-4AA9-98E7-23B707533026}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FF5C2AA7-0727-4AA9-98E7-23B707533026}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FF5C2AA7-0727-4AA9-98E7-23B707533026}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = Owin.StatelessAuthExample\Owin.StatelessAuthExample.csproj
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
