+++ src/Owin.StatelessAuth/Properties/AssemblyInfo.cs
