--- src/Owin.StatelessAuthExample.Tests/App.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
-    </startup>
-  <appSettings>
-    <add key="securekey" value="123"/>
-  </appSettings>
-</configuration>
