--- src/Owin.StatelessAuthExample.Tests/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Owin" version="3.0.0-rc2" targetFramework="net45" />
-  <package id="Microsoft.Owin.Hosting" version="3.0.0-rc2" targetFramework="net45" />
-  <package id="Microsoft.Owin.Testing" version="3.0.0-rc2" targetFramework="net45" />
-  <package id="Owin" version="1.0" targetFramework="net45" />
-  <package id="xunit" version="1.9.2" targetFramework="net45" />
-</packages>
