--- src/Owin.StatelessAuthExample/DemoUserIdentity.cs
@@ -1,11 +1,0 @@
-namespace Owin.StatelessAuthExample
-{
-    using System.Collections.Generic;
-    using Nancy.Security;
-
-    public class DemoUserIdentity : IUserIdentity
-    {
-        public string UserName { get; set; }
-        public IEnumerable<string> Claims { get; set; }
-    }
-}
