--- src/Owin.StatelessAuthExample/IConfigProvider.cs
@@ -1,8 +1,0 @@
-namespace Owin.StatelessAuthExample
-{
-    public interface IConfigProvider
-    {
-        string GetAppSetting(string propertyName);
-        T GetAppSetting<T>(string propertyName) where T : struct;
-    }
-}
