--- src/Owin.StatelessAuthExample/IJwtWrapper.cs
@@ -1,9 +1,0 @@
-namespace Owin.StatelessAuthExample
-{
-    using JWT;
-
-    public interface IJwtWrapper
-    {
-        string Encode(object payload, string key, JwtHashAlgorithm algorithm);
-    }
-}
