--- src/Owin.StatelessAuthExample/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="JWT" version="1.3.2" targetFramework="net45" />
-  <package id="Microsoft.Owin" version="2.1.0" targetFramework="net45" />
-  <package id="Microsoft.Owin.Host.HttpListener" version="2.1.0" targetFramework="net45" />
-  <package id="Microsoft.Owin.Hosting" version="2.1.0" targetFramework="net45" />
-  <package id="Nancy" version="0.22.2" targetFramework="net45" />
-  <package id="Nancy.Owin" version="0.22.2" targetFramework="net45" />
-  <package id="Owin" version="1.0" targetFramework="net45" />
-</packages>
