+++ test/Owin.StatelessAuth.Jwt.Tests/packages.config
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net45" />
+  <package id="Microsoft.Owin.Hosting" version="3.0.1" targetFramework="net45" />
+  <package id="Microsoft.Owin.Testing" version="3.0.1" targetFramework="net45" />
+  <package id="Owin" version="1.0" targetFramework="net45" />
+  <package id="xunit" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.abstractions" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.assert" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.core" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.extensibility.core" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.runner.visualstudio" version="2.0.0-rc4-build1049" targetFramework="net45" />
+</packages>
