+++ test/Owin.StatelessAuth.Tests/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="FakeItEasy" version="1.19.0" targetFramework="net45" />
+  <package id="xunit" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.abstractions" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.assert" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.core" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.extensibility.core" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.extensions" version="2.0.0-rc4-build2924" targetFramework="net45" />
+  <package id="xunit.runner.visualstudio" version="2.0.0-rc4-build1049" targetFramework="net45" />
+</packages>
